Add abstraction to Day 18
