+++ Day18 - Many-Worlds Interpretation/Day18.csproj
@@ -4,6 +4,11 @@
     <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
     <Nullable>enable</Nullable>
+    <RootNamespace>AdventOfCode.Year2019.Day18</RootNamespace>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="mMosiur.AdventOfCode.Abstractions" Version="1.0.1" />
+  </ItemGroup>
 
 </Project>
