+++ Day18 - Many-Worlds Interpretation/Nodes/DoorNode.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Day18;
+namespace AdventOfCode.Year2019.Day18;
 
 public class DoorNode : INode
 {
