+++ Day18 - Many-Worlds Interpretation/Nodes/Node.cs
@@ -1,6 +1,6 @@
 using System.IO;
 
-namespace Day18;
+namespace AdventOfCode.Year2019.Day18;
 
 public static class Node
 {
