+++ Day18 - Many-Worlds Interpretation/TunnelMap/TunnelMap.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Day18;
+namespace AdventOfCode.Year2019.Day18;
 
 public class TunnelMap
 {
