+++ Day18 - Many-Worlds Interpretation/TunnelMap/TunnelMapTraverser.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Linq;
 
-namespace Day18;
+namespace AdventOfCode.Year2019.Day18;
 
 public class TunnelMapRobots
 {
