+++ Day18 - Many-Worlds Interpretation/Utils/ArrayExtensions.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Linq;
-namespace Day18;
+namespace AdventOfCode.Year2019.Day18;
 
 public static class ArrayExtensions
 {
