+++ Day18 - Many-Worlds Interpretation/Utils/Point.cs
@@ -1,3 +1,3 @@
-namespace Day18;
+namespace AdventOfCode.Year2019.Day18;
 
 public record struct Point(int X, int Y);
